# static/openapi.yaml
openapi: 3.0.0
info:
  title: ChatGPT Song Player API
  version: 1.0.0
  description: A simple API to queue and play songs using yt-dlp and mpv.

servers:
  - url: http://localhost:5000

paths:
  /queue_songs:
    post:
      operationId: queueSongs
      summary: Add a list of songs to the playback queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - song_names
              properties:
                song_names:
                  type: array
                  items:
                    type: string
                  description: List of the songs to be queued
      responses:
        '200':
          description: Songs added to queue successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Song names not provided.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /play_songs:
    post:
      operationId: playSongs
      summary: Clear the queue and play the provided songs in order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - song_names
              properties:
                song_names:
                  type: array
                  items:
                    type: string
                  description: List of the songs to be played immediately.
      responses:
        '200':
          description: Songs will play now.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Song names not provided.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /skip_songs:
    post:
      operationId: skipSongs
      summary: Skips a number of songs, given by the request body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - number
              properties:
                number:
                  type: integer
                  description: Number of songs to skip
      responses:
        '200':
          description: Songs have been skipped.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid number given
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /stop_playback:
    post:
      operationId: stopPlayback
      summary: Stop playback
      responses:
        '200':
          description: Playback stopped successfully
          content: 
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /start_playback:
    post:
      operationId: startPlayback
      summary: Start playback
      responses:
        '200':
          description: Playback started successfully
          content: 
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /clear_queue:
    post:
      operationId: clearQueue
      summary: Clear the song queue.
      responses:
        '200':
          description: Queue cleared successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string